output updates
This module is still in the early phases. Added unicode characters to
the base Terminal class. In the near future this class will be used to
store information about the users terminal to provide better and clearer
output depending on the situation. EG if a terminal doesn't provide
unicode character support, this class will be able to tell.

Adds super inheritance to the Outputs class. This inherits from Terminal
base class. Depending on the symlinks status this class also adjusts the
symbols that are printed to the user. WIP